refactor(core): isolate environment partition refresh
